requirements.txt
Flask
Pillow
nltk
werkzeug
Flask-Cors
gunicorn
cloudinary
requests